Reviewed-on: https://git.smartdata.solutions/smartdata/smartskills/pulls/127

+++ src/components/SkillItem/SkillItem.tsx
@@ -3,6 +3,17 @@ import { DotsVerticalIcon } from "@radix-ui/react-icons";
 import React from "react";
 
 import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+import {
   Dialog,
   DialogContent,
   DialogDescription,
@@ -51,43 +62,65 @@ const SkillItem = ({ skill, isLoggedUser }: SkillItemProps) => {
 
   return (
     <div className="flex flex-col mb-4">
-      <div className="flex justify-between items-start mb-2">
+      <div className="flex justify-between items-start gap-2 mb-2">
         <p className="font-medium text-slate-700">{skill.name}</p>
 
         <div className="flex items-center gap-4">
           <p className="font-medium text-slate-700">{skill.rating}%</p>
           {isLoggedUser && (
             <Dialog>
-              <DropdownMenu>
-                <DropdownMenuTrigger>
-                  <DotsVerticalIcon className="text-slate-600 mb-[0.1rem]" />
-                </DropdownMenuTrigger>
-                <DropdownMenuContent>
-                  <DropdownMenuItem asChild>
-                    <DialogTrigger className="w-full text-md font-medium text-slate-500">
-                      Edit
-                    </DialogTrigger>
-                  </DropdownMenuItem>
-                  <DropdownMenuSeparator />
-                  <DropdownMenuItem
-                    className="text-red-500"
-                    onClick={handleDeleteFromUser(skill.id, skill.name)}
-                  >
-                    Delete
-                  </DropdownMenuItem>
-                </DropdownMenuContent>
-              </DropdownMenu>
+              <AlertDialog>
+                <DropdownMenu>
+                  <DropdownMenuTrigger>
+                    <DotsVerticalIcon className="text-slate-600" />
+                  </DropdownMenuTrigger>
+                  <DropdownMenuContent className="w-60 mr-2">
+                    <DropdownMenuItem asChild>
+                      <DialogTrigger className="w-full text-md font-medium text-slate-500">
+                        Edit
+                      </DialogTrigger>
+                    </DropdownMenuItem>
+                    <DropdownMenuSeparator />
 
-              <DialogContent>
-                <DialogHeader>
-                  <DialogTitle>Skill</DialogTitle>
-                  <DialogDescription>
-                    Edit skill knowledge level.
-                  </DialogDescription>
-                </DialogHeader>
+                    <DropdownMenuItem className="text-red-500" asChild>
+                      <AlertDialogTrigger asChild>
+                        <p>Delete {`${skill.name}`}</p>
+                      </AlertDialogTrigger>
+                    </DropdownMenuItem>
+                  </DropdownMenuContent>
+                </DropdownMenu>
 
-                <EditRating skill={skill} />
-              </DialogContent>
+                <AlertDialogContent>
+                  <AlertDialogHeader>
+                    <AlertDialogTitle>
+                      Are you absolutely sure?
+                    </AlertDialogTitle>
+                    <AlertDialogDescription>
+                      This action cannot be undone.
+                    </AlertDialogDescription>
+                  </AlertDialogHeader>
+                  <AlertDialogFooter>
+                    <AlertDialogCancel>Cancel</AlertDialogCancel>
+                    <AlertDialogAction
+                      onClick={handleDeleteFromUser(skill.id, skill.name)}
+                      className="bg-destructive text-destructive-foreground shadow-sm hover:bg-destructive/90"
+                    >
+                      Delete
+                    </AlertDialogAction>
+                  </AlertDialogFooter>
+                </AlertDialogContent>
+
+                <DialogContent>
+                  <DialogHeader>
+                    <DialogTitle>Skill</DialogTitle>
+                    <DialogDescription>
+                      Edit skill knowledge level.
+                    </DialogDescription>
+                  </DialogHeader>
+
+                  <EditRating skill={skill} />
+                </DialogContent>
+              </AlertDialog>
             </Dialog>
           )}
         </div>
